About 1.1
Portrait added to the About section.

Background image (the same one as the Projects section) has been added to the About section.

Button colour and button colour on hover has been changed for harmony with other sections.

Background colour have been added to the BioContainer to make the text visible in contrast with the new background image.

[Social Section]
A header has been added to the Social section.