About 1.2
Bio has been inserted.

[Projects Section]
Buttons have been added to the each project. One for the code on GitHub and the other for the live example.

The first project (Seneca Online Bookstore) has been added to its respective place with an image, a description and GitHub link.

An outline added for each project around their divs.