Sidebar 1.2
Sidebar toggler now changes color and background-color on hover.

Doesn't have a buzz-out effect on hover anymore.

There's a slight glitch with its changing color. For a split second it changes background to none and then to the new color on hover. Noticeable when paid attention but not a huge deal. Should be sorted out before the final production if possible.